+++ packages/kiarina-currency/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `get_system_currency()` function for automatic system currency detection from locale settings
+
 ## [1.28.0] - 2026-01-16
 
 ### Added